simulator/requirements.txt
paho-mqtt==1.6.1
requests==2.31.0
Pillow==10.1.0
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
